requirements.txt
asgiref==3.5.0
boto3==1.21.37
botocore==1.24.37
click==8.1.2
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
Flask==2.1.1
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
pipdeptree==2.2.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
